requirements.txt
SQLAlchemy~=2.0.40
Flask-Migrate~=4.1.0
fastapi~=0.115.12
alembic~=1.16.1
uvicorn~=0.34.2
langchain~=0.3.25
langchain-community>=0.0.10
langchain-core>=0.1.0
openai>=1.0.0
pydantic~=2.11.5
langchain-openai~=0.3.18
python-dotenv~=1.1.0
pytest>=7.0
scikit-learn~=1.6.1
nltk~=3.9.1